moved more functions out of config_metplus
