+++ internal/tests/pytests/util/config_metplus/test_config_metplus.py
@@ -913,65 +913,3 @@ def test_format_var_items_options_semicolon(config_value,
     var_items = config_metplus._format_var_items(field_configs, time_info)
     result = var_items.get('extra')
     assert result == expected_result
-
-
-@pytest.mark.parametrize(
-    'input_dict, expected_list', [
-        ({'init': datetime(2019, 2, 1, 6),
-          'lead': 7200, },
-         [
-             {'index': '1',
-              'fcst_name': 'FNAME_2019',
-              'fcst_level': 'Z06',
-              'obs_name': 'ONAME_2019',
-              'obs_level': 'L06',
-             },
-             {'index': '1',
-              'fcst_name': 'FNAME_2019',
-              'fcst_level': 'Z08',
-              'obs_name': 'ONAME_2019',
-              'obs_level': 'L08',
-             },
-         ]),
-        ({'init': datetime(2021, 4, 13, 9),
-          'lead': 10800, },
-         [
-             {'index': '1',
-              'fcst_name': 'FNAME_2021',
-              'fcst_level': 'Z09',
-              'obs_name': 'ONAME_2021',
-              'obs_level': 'L09',
-              },
-             {'index': '1',
-              'fcst_name': 'FNAME_2021',
-              'fcst_level': 'Z12',
-              'obs_name': 'ONAME_2021',
-              'obs_level': 'L12',
-              },
-         ]),
-    ]
-)
-@pytest.mark.util
-def test_sub_var_list(metplus_config, input_dict, expected_list):
-    config = metplus_config
-    config.set('config', 'FCST_VAR1_NAME', 'FNAME_{init?fmt=%Y}')
-    config.set('config', 'FCST_VAR1_LEVELS', 'Z{init?fmt=%H}, Z{valid?fmt=%H}')
-    config.set('config', 'OBS_VAR1_NAME', 'ONAME_{init?fmt=%Y}')
-    config.set('config', 'OBS_VAR1_LEVELS', 'L{init?fmt=%H}, L{valid?fmt=%H}')
-
-    time_info = ti_calculate(input_dict)
-
-    actual_temp = config_metplus.parse_var_list(config)
-
-    pp = pprint.PrettyPrinter()
-    print(f'Actual var list (before sub):')
-    pp.pprint(actual_temp)
-
-    actual_list = config_metplus.sub_var_list(actual_temp, time_info)
-    print(f'Actual var list (after sub):')
-    pp.pprint(actual_list)
-
-    assert len(actual_list) == len(expected_list)
-    for actual, expected in zip(actual_list, expected_list):
-        for key, value in expected.items():
-            assert actual.get(key) == value
