+++ metplus/util/run_util.py
@@ -6,10 +6,11 @@ from importlib import import_module
 
 from .constants import NO_COMMAND_WRAPPERS
 from .system_util import get_user_info, write_list_to_file
-from .config_util import get_process_list
+from .config_util import get_process_list, handle_env_var_config
+from .config_util import handle_tmp_dir, write_final_conf, write_all_commands
 from .config_validate import validate_config_variables
 from .. import get_metplus_version
-from . import config_metplus
+from .config_metplus import setup
 from . import camel_to_underscore
 
 
@@ -19,7 +20,7 @@ def pre_run_setup(config_inputs):
     print(f'Starting METplus v{version_number}')
 
     # Read config inputs and return a config instance
-    config = config_metplus.setup(config_inputs)
+    config = setup(config_inputs)
 
     logger = config.logger
 
@@ -67,10 +68,10 @@ def pre_run_setup(config_inputs):
                    os.path.join(config.getdir('OUTPUT_BASE'), "stage"))
 
     # handle dir to write temporary files
-    config_metplus.handle_tmp_dir(config)
+    handle_tmp_dir(config)
 
     # handle OMP_NUM_THREADS environment variable
-    config_metplus.handle_env_var_config(config,
+    handle_env_var_config(config,
                           env_var_name='OMP_NUM_THREADS',
                           config_name='OMP_NUM_THREADS')
 
@@ -136,7 +137,7 @@ def run_metplus(config):
         # if process list contains any wrapper that should run commands
         if any([item[0] not in NO_COMMAND_WRAPPERS for item in process_list]):
             # write out all commands and environment variables to file
-            if not config_metplus.write_all_commands(all_commands, config):
+            if not write_all_commands(all_commands, config):
                 # report an error if no commands were generated
                 total_errors += 1
 
@@ -179,7 +180,7 @@ def post_run_cleanup(config, app_name, total_errors):
                                          '%Y%m%d%H%M%S')
 
     # rewrite final conf so it contains all of the default values used
-    config_metplus.write_final_conf(config)
+    write_final_conf(config)
 
     # compute time it took to run
     end_clock_time = datetime.now()
